// src/app/modules/variable/core/models/variable.enums.ts
export enum EVariableType {
  SELECTBOX = "selectbox",
  CHECKBOX = "checkbox",
  MULTIPLE_CHECKBOX = "multiple_checkbox",
  MULTIPLE_SELECTBOX = "multiple_selectbox",
}

export enum EVariableTypeL {
  selectbox = "Seçim Kutusu",
  checkbox = "Onay Kutusu",
  multiple_checkbox = "Çoklu Onay Kutusu",
  multiple_selectbox = "Çoklu Seçim Kutusu",
}